* JAVA FEATURE: Added most features of com.madara.threads.BaseThread and com.madara.threads.Threader   * Features are untested   * Threader is not currently setting up terminated and paused variables inside of BaseThread   * Will start testing threading soon * FEATURE: Added Madara::Utility logging options such as Madara::Utility::set_log_level and Madara::Utility::log for simple logging.   * Changed MADARA_debug_level to be an int instead of an unsigned int to get rid of warnings about unsigned comparisons for 0 * Updated to version 2.0.3

+++ include/madara/knowledge_engine/Knowledge_Base_Impl.inl
@@ -205,7 +205,7 @@ inline void
 Madara::Knowledge_Engine::Knowledge_Base_Impl::print (
   unsigned int level) const
 {
-  MADARA_DEBUG (level, (LM_INFO, 
+  MADARA_DEBUG ((int)level, (LM_INFO, 
     "\nKnowledge in Knowledge Base:\n"));
   map_.print (level);
 }
